refactor: Consolidate duplicate MockStorageManager implementations
**Feedback:** Duplicate MockStorageManager implementations across test files create maintenance overhead and potential inconsistencies between test implementations
**Solution:** Created shared mock in internal/testutil/mocks.go with best features from both implementations, updated all test files to use shared implementation with helper methods
**Rationale:** Reduces code duplication, ensures consistency across tests, and centralizes mock maintenance in one location

Comment ID: 2217362377
Review Comment: https://github.com/biwakonbu/ai-pr-review-checker/pull/4#discussion_r2217362377
